requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
chromadb>=0.4.0
langchain>=0.0.350
langchain-openai>=0.0.2
python-multipart>=0.0.6
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.1.0 